+++ Examples/SlackClone/SlackReactionsView.swift
@@ -6,6 +6,19 @@ import StreamChat
 import StreamChatUI
 import UIKit
 
+extension MessageReactionType {
+    func toEmoji() -> String {
+        let emojis: [String: String] = [
+            "love": "❤️",
+            "haha": "😂",
+            "like": "👍",
+            "sad": "😔",
+            "wow": "🤯"
+        ]
+        return emojis[rawValue] ?? "❓"
+    }
+}
+
 final class SlackReactionsView: _View, ThemeProvider, UICollectionViewDataSource, UICollectionViewDelegate {
     var content: ChatMessage? {
         didSet { updateContent() }
@@ -25,8 +38,10 @@ final class SlackReactionsView: _View, ThemeProvider, UICollectionViewDataSource
 
     let reactionWidth: CGFloat = 40
     let reactionRowHeight: CGFloat = 30
+    let reactionsMarginLeft: CGFloat = 40
+    let reactionInterSpacing: CGFloat = 4
 
-    private lazy var collectionView: UICollectionView = {
+    lazy var collectionView: UICollectionView = {
         let collectionView = UICollectionView(frame: .zero, collectionViewLayout: collectionViewLayout)
         collectionView.translatesAutoresizingMaskIntoConstraints = false
         collectionView.dataSource = self
@@ -48,9 +63,9 @@ final class SlackReactionsView: _View, ThemeProvider, UICollectionViewDataSource
             ),
             subitems: [item]
         )
-        group.interItemSpacing = .fixed(4)
+        group.interItemSpacing = .fixed(reactionInterSpacing)
         let section = NSCollectionLayoutSection(group: group)
-        section.interGroupSpacing = 4
+        section.interGroupSpacing = reactionInterSpacing
 
         return UICollectionViewCompositionalLayout(section: section)
     }
@@ -62,8 +77,8 @@ final class SlackReactionsView: _View, ThemeProvider, UICollectionViewDataSource
 
         addSubview(collectionView)
         NSLayoutConstraint.activate([
-            collectionView.topAnchor.constraint(equalTo: topAnchor),
-            collectionView.leadingAnchor.constraint(equalTo: leadingAnchor),
+            collectionView.topAnchor.constraint(equalTo: topAnchor, constant: -4),
+            collectionView.leadingAnchor.constraint(equalTo: leadingAnchor, constant: reactionsMarginLeft),
             collectionView.trailingAnchor.constraint(equalTo: trailingAnchor),
             collectionView.bottomAnchor.constraint(equalTo: bottomAnchor)
         ])
@@ -82,7 +97,8 @@ final class SlackReactionsView: _View, ThemeProvider, UICollectionViewDataSource
 
         collectionView.reloadData()
 
-        let numberOfRows = Double(reactions.count) * reactionWidth / UIScreen.main.bounds.width
+        let screenWidth = UIScreen.main.bounds.width
+        let numberOfRows = Double(reactions.count) * (reactionWidth + reactionInterSpacing) / (screenWidth - reactionsMarginLeft)
         heightConstraint?.constant = ceil(numberOfRows) * reactionRowHeight
     }